requirements.txt
streamlit>=1.36.0
openai>=1.40.0
python-dotenv>=1.0.1

# PDF parsing
pypdf>=4.0.0

# Math + arrays
numpy>=1.26.0

# Chroma vector store
chromadb>=0.5.0

# LangChain framework
langchain>=0.2.0

# LangChain community package for integrations
langchain-community>=0.2.0

